pyproject.toml
[tool.poetry]
name = "traefik2piholedns"
version = "0.1.0"
description = ""
authors = ["meska <meskatech@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
paramiko = "^3.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
